Fixes data type

+++ db/migrate/20240408233625_create_generative_text_presets.rb
@@ -4,7 +4,7 @@ class CreateGenerativeTextPresets < ActiveRecord::Migration[7.1]
       t.string :name, null: false
       t.string :description, null: false
       t.text :system_message, null: false
-      t.integer :temperature, null: false
+      t.float :temperature, null: false
 
       t.timestamps
     end
